Move API key to nuxt.config.
