+++ apps/release-notes/.env.example
@@ -1,1 +1,2 @@
 ZENHUB_APIKEY
+ENTITIES_GHID
