+++ apps/release-notes/nuxt.config.ts
@@ -1,15 +1,22 @@
 // https://nuxt.com/docs/api/configuration/nuxt-config
 export default defineNuxtConfig({
   devtools: { enabled: true },
-  modules: [
-    '@nuxtjs/tailwindcss'
-  ],
+  modules: ["@nuxtjs/tailwindcss"],
   tailwindcss: {
-    cssPath: '~/assets/css/tailwind.css',
-    configPath: 'tailwind.config'
+    cssPath: "~/assets/css/tailwind.css",
+    configPath: "tailwind.config",
     // exposeConfig: false,
     // config: {},
     // injectPosition: 0,
     // viewer: true,
-  }
-})
+  },
+  runtimeConfig: {
+    // Private keys are only available on the server
+    // apiSecret: '123',
+
+    // Public keys that are exposed to the client
+    public: {
+      zenhubAPI: process.env.ZENHUB_APIKEY,
+    },
+  },
+});
